@@ -4,13 +4,16 @@
 	"description": "Get namespaced config from the closest package.json",
 	"license": "MIT",
 	"repository": "sindresorhus/pkg-conf",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=6"
+		"node": "^12.20.0 || ^14.13.1 || >=16.0.0"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -27,23 +30,21 @@
 		"fs",
 		"graceful",
 		"load",
-		"pkg",
 		"package",
 		"config",
-		"conf",
 		"configuration",
 		"object",
 		"namespace",
 		"namespaced"
 	],
 	"dependencies": {
-		"find-up": "^3.0.0",
-		"load-json-file": "^5.2.0"
+		"find-up": "^6.0.0",
+		"load-json-file": "^7.0.0"
 	},
 	"devDependencies": {
-		"ava": "^1.4.1",
-		"tsd": "^0.7.2",
-		"xo": "^0.24.0"
+		"ava": "^3.15.0",
+		"tsd": "^0.17.0",
+		"xo": "^0.44.0"
 	},
 	"fixture": {
 		"foo": true
